pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pulsecast"
version = "0.1.0"
description = "Probabilistic shipment demand forecasting using NYC TLC trip records and live MTA GTFS-Realtime congestion signals."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12,<3.14"
authors = [
  { name = "Paulo Olveira" },
]
keywords = ["forecasting", "demand", "nyc", "tlc", "gtfs", "timeseries"]

dependencies = [
  # Data / storage
  "polars>=0.20.0",
  "pyarrow>=14.0.0",
  "psycopg2-binary>=2.9.9",

  # Ingestion
  "requests>=2.31.0",
  "types-requests>=2.31.0",
  "gtfs-realtime-bindings>=1.0.0",
  "apscheduler>=3.10.4",
  "boto3>=1.34.0",

  # Feature store

  # Models
  "lightgbm>=4.3.0",
  "statsforecast>=1.7.0",
  "pytorch-forecasting>=1.1.0",
  "torch>=2.2.0",
  "pytorch-lightning>=2.2.0",
  "scikit-learn>=1.4.0",

  # ONNX
  "onnxmltools>=1.16.0",
  "onnxruntime>=1.17.0",

  # Serving
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.6.0",
  "redis>=5.0.0",

  # Dashboard
  "streamlit>=1.32.0",
  "plotly>=5.20.0",

  # Experiment tracking / monitoring
  "mlflow>=2.12.0",
  "evidently>=0.4.22",

  # Calendar features
  "holidays>=0.47",

  # geo
  "pandas>=2.2.0",
  "geopandas>=1.0.0",
  "pyogrio>=0.9.0",
  "shapely>=2.0.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.27.0",
  "ruff>=0.4.0",
  "mypy>=1.9.0",
  "nbformat>=4.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["pulsecast"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "kaleido (>=1.2.0,<2.0.0)",
    "plotly (>=6.6.0,<7.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)"
]
